GemRequirementTest: Default to check gem rdoc
The gem `ruby` was previously used to check that `gem`
was able to determine that a gem was installed.
The gem `ruby` is not packaged as part of `ruby` on
many distributions, and Fedora `ruby` has no gems included.
However, one of the gems required by `gem` is `rdoc`.

Fixes https://gitlab.com/coala/package_manager/issues/49