DistributionRequirement: Implement default package
Allow a standard `package` name to be used for all distribution
package managers, with overrides for specific managers that
use a different package name.

`DistributionRequirement.package` is now a string, complying with
the expected behaviour of PackageRequirement classes, using
the `package` argument as the default if set, falling back to the
package name for the hosts preferred package manager, and then
falling back to the most common override.

Also add unimplemented `version` and `repo` to ensure constructor
signature matches the base PackageRequirement.

Fixes https://gitlab.com/coala/package_manager/issues/16
Closes https://gitlab.com/coala/package_manager/issues/63