DistributionRequirement: Fix REQUIREMENTS
be1d082 contained a logic error, which would result in
assuming a supported package manager was present when
only `grep` was present.

Only one package manager is required, and grep is always
required as that is the only currently possible way to
record the dependency which exists for only a few of the
package managers.

Fixes https://gitlab.com/coala/package_manager/issues/108
Related to https://gitlab.com/coala/package_manager/issues/75