.gitlab-ci.yml: Avoid pre-release of coala-bears
ApertiumLintBear in the bears 0.12 pre-release requires many
host library and utilities, via package libxslt1-dev or similar.
As it is not used, it is more efficient to remove it, and
the hard dependency should be removed before 0.12 release
due to https://github.com/coala/coala-bears/issues/1000 .

Also undo the VintBear workaround, which is unnecessary after 0.11.1.